Reduce number of allocations
Rather than duplicating every line from the input into the choices list,
maintain a single large buffer and represent the choices as ranges in
that buffer. This means the only memory allocations that occur are when
we need to expand the buffer. For small inputs, this means we end up
with only a single allocation.